This fixes a security issue.
If you hack a filename into something stupid you can get the server to run commands.  Not sure how serious this is since a lot of my testing failed..... 2 things that did manage to run were touch, and rm. (with rm maybe being something of a big deal?).... The rm was (only?) successful in my share that has write access for "everyone" so that might be related.  The bigger issue is more you can trick the download plugin to run code, vs what you can do with that code.